refactor(ATT-473): extract boot diagnostics into application_bootdiag.cpp
Follow the modular split convention (ATT-424) instead of growing
application.cpp. Boot-record read/classify, periodic NVS snapshot, and the
reset-reason helper move into a dedicated partial translation unit; the
class declaration stays in application.hpp and the call sites remain in
application.cpp.
